Redesignate findProductWithId() (singular); add findProductsWithId() (plural)
The findProductWithId() method is  now getProductWithId() to indicate that the specified ID is assumed to exist. This is used for the path param URL requests. Alternatively, the findProductsWithId() method was created to support queries in which the specified ID is not assumed to exist. This is used for query param URL requests.